added cours policies

+++ app/Http/Controllers/CoursController.php
@@ -38,6 +38,7 @@ class CoursController extends Controller
     {
         $user = Auth::user();
         $courses = Cours::get();
+        $this->authorize('subscriptionlist', $courses[0]);
         $my_courses = null;
         foreach($courses as $course){
             foreach($course->user as $c_user)
@@ -56,7 +57,9 @@ class CoursController extends Controller
     function my_formation()
     {
         $user = Auth::user();
+        
         $cours = Cours::where('formation_id', '=', $user->formation_id)->get();
+        $this->authorize('myformationlist', $cours[0]);
         return view('coursList', ['cours' => $cours]);
     }
 
@@ -65,6 +68,7 @@ class CoursController extends Controller
     {
         $user = Auth::user();
         $cours = Cours::where('user_id', '=', $user->id)->get();
+        $this->authorize('iteach', $cours[0]);
         return view('coursList', ['cours' => $cours]);
     }
 
@@ -84,7 +88,11 @@ class CoursController extends Controller
     function subscribe(Request $request, $cours_id)
     {
         $user = Auth::user();
+
         $cours = Cours::findOrFail($cours_id);
+
+        $this->authorize('subscribe', $cours);
+
         $cours->user()->attach($user);
 
         $request->session()->flash('etat', 'Inscription au cours effectué !');
@@ -95,7 +103,11 @@ class CoursController extends Controller
     function unsubscribe(Request $request, $cours_id)
     {
         $user = Auth::user();
+
         $cours = Cours::findOrFail($cours_id);
+
+        $this->authorize('unsubscribe', $cours);
+
         $cours->user()->detach($user);
 
         $request->session()->flash('etat', 'Désinscription au cours effectué !');
